Add search result highlight to search result highlight display
Fix #70

This is for sitewide search, not the news search which uses the teaser display
mode.

Usual cavet about using search_api as dependency applies.
(If not using search_api, the search result display will be blank)

Note: Whilst a newsroom search result display mode is included, there is no
search index for newsroom as there is no content to search.