Update the news list view to use the newsroom nid filter
In order to get the correct title for the feed, it's nescearry to
change the argument filter from numeric to nid, this means also
adding a relationship to the newsroom node and using the id
of that as the argument, instead of the field
localgov_newsroom target_id.

Also update the feed machine name to rss_feed from feed_1.